Commit before evaluating quantized model

+++ src/utils/model_conversion_utils.py
@@ -29,8 +29,8 @@ class SingleOutputRegressorProxy:
     def __getattr__(self, name):
         return getattr(self._model, name)
 
-def convert_model(model, save_dir):
+def convert_model(model, save_dir, use_dtype='float'):
     for i, col in enumerate(['qw', 'qx', 'qy', 'qz']):
         proxy = SingleOutputRegressorProxy(model, i)
-        cmodel = emlearn.convert(proxy, kind='DecisionTreeRegressor', method='inline', dtype='float')
+        cmodel = emlearn.convert(proxy, kind='DecisionTreeRegressor', method='inline', dtype=use_dtype)
         cmodel.save(name=f'model_{col}', file=f'{save_dir}/model_{col}.h')
